pythonProject/requirement.txt
Flask
openai
request